+++ srcs/parsing/build_ast.c
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/16 11:47:44 by vviterbo          #+#    #+#             */
-/*   Updated: 2025/04/17 19:49:26 by vviterbo         ###   ########.fr       */
+/*   Updated: 2025/04/17 23:34:43 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
