+++ srcs/parsing/expand.c
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/07 11:51:06 by vviterbo          #+#    #+#             */
-/*   Updated: 2025/04/17 19:49:31 by vviterbo         ###   ########.fr       */
+/*   Updated: 2025/04/17 22:58:30 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
