+++ srcs/parsing/parse.c
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/12 14:45:49 by vviterbo          #+#    #+#             */
-/*   Updated: 2025/04/15 15:02:40 by vviterbo         ###   ########.fr       */
+/*   Updated: 2025/04/17 23:42:28 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	**parse_str(t_data *data, char *str, t_token_type type)
 
 	isescaped = is_quote_escaped(data, str);
 	if (data->exit_status)
-		return (NULL);	
+		return (NULL);
 	parsed = expand_var(data, str, isescaped);
 	if (data->exit_status)
 		return (NULL);
@@ -33,4 +33,3 @@ char	**parse_str(t_data *data, char *str, t_token_type type)
 	free(isescaped);
 	return (parsed_arr);
 }
-
