+++ srcs/utils/token_utils.c
@@ -6,7 +6,7 @@
 /*   By: vviterbo <vviterbo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/09 12:26:45 by vviterbo          #+#    #+#             */
-/*   Updated: 2025/04/17 19:50:05 by vviterbo         ###   ########.fr       */
+/*   Updated: 2025/04/17 23:35:11 by vviterbo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
